Add federal query parameter to the API
Such that we can return only holidays for which federal=true or federal=false.

Note: removed the more efficient "get ProvinceHoliday where {}id = ?" query
because at this point it's really not giving us a lot, it just makes the
code more confusing.

If we wanted it back, we should pass in the list of holiday/province ids,
because that would work.