Add nextHoliday method for front page
It's a bit ugly, but unfortuntately we can't really get this into
the queries file the way it's currently written.

Queries.js always returns ana rray of stuff. What we want to return
in this case is the next holiday for all of Canada, which,
unfortunately, doesn't follow the pattern we have or even conventional
understandings of what is next.

Since there are several holidays on the same days, this new method
- finds the "next" date in a list of holidays
- finds all the holidays which match the "next" date
- returns the holiday with the highest number of provinces