Take out linebreak in favour of 80% width
We want the date to be pretty obvious, which means it should
sit on its own if possible.
But we also don't want it breaking so that the word "on" is all
by itself.

Decided to go with an 80% width, which means it breaks pretty early
but not all the time.