Set initial section to the innerheight of the window on the first load
For reasons of elegance or design or whatever, ios safari (and
now chrome on android) will initially show larger top and bottom
browser interface that shrinks as you scroll down the page.

Unfortunately, the "vh" unit doesn't take this expanding/shrinking
browser into account so what ends up happening is that things you want
stuck to the bottom of the interface are actually hidden by default.
(unless you use fixed positioning).

The way to solve this is some good ol' javascript that sets a CSS
variable (pretty cool, right?), and, yes, this literally is currently
the best fix going.

Sources:
- https://nicolas-hoizey.com/2015/02/viewport-height-is-taller-than-the-visible-part-of-the-document-in-some-mobile-browsers.html
- https://medium.com/@susiekim9/how-to-compensate-for-the-ios-viewport-unit-bug-46e78d54af0d
- https://css-tricks.com/the-trick-to-viewport-units-on-mobile/