Correct function for finding relative dates
What was happening was that the string said

"Monday before May 25"

And my algorithm will find the Date for May 25, and then count
backwards 1 day at a time until it found the right weekday.

However, if May 25 _was_ a Monday, then it wouldn't ever count
backwards: it would just return May 25.

Preincrementing the counter resolves this, and doesn't break any of
the existing tests 👍🙌