POST route for provinces
If we use the select and click the submit button, it should
redirect us where we want to go. :)

Thought about putting in a whitelist but I don't really think it's
that important.
I limit province IDs to 2 characters and then uppercase them before
encoding the URL. Shouldn't really be a way to get anything fishy in
there, and if you put in some bad characters, you'll just get a 404.