add amazing logic to NextHolidayBox
not my finest moment but the job is getting done.

Pretty much, the process is that I duplicated a bunch of code before
so that I could see what I was aiming for, and now I am slowly refactoring
code and stuff so that I can generalize the base components to be
more flexible in terms of what gets passed in.

In this specific case, we're looking at 'NextHolidayBox', which
is pretty specifically named assuming there's a next holiday.

For other years though, there won't be a next holiday, so we want to
have a variant where it just says the province name at the top and
then the year.

Pretty sure all the refaccing is done, as well as added a bunch of tests..