removed a bunch of useless /ics routes
basically, instead of removing the current year and redirecting,
just return the results. the download link isn't really a thing
you type into a url, so it doesn't matter at all if there's a
shortcut to get the current year.

less code == less logic == simpler, better, etc

bumped the version to 2!!! as this is actually a breaking change.
If somehow you loved the old URLs, they are all gone now.
Risk of this is pretty much nothing.