add express-openapi-validator and get it working
had a few false starts with this.
basically, it's hard to see if it's really working when everything
just responds as usual.
in the end, was able to verify it was working by trying to break it.

this was pretty messy code but basically I installed the new packaged,
followed their README, and got it to work.

source:
- https://www.npmjs.com/package/express-openapi-validator