overhaul API routes
added in the OpenAPIValidator and am testing both requests and responses.

lots of heavy lifting here, and, to be honest, most of my error
messages got worse, but at the end of the day we have a SPEC.

the main thing that happened is that I split off the "/api"
route, which is actually a page that we load, off from the "v1"
routes, which are the actual API.

since the "v1" routes may change in the future, and since they should be
treated differently that pages in the app, I created a couple of routers
where before they were just one.

the "api" router is more general and uses the "v1" router.
Similarly, we could make a "v2" router one day and it would all be chill.

Found this one here pretty helpful:
- https://gist.github.com/zcaceres/f38b208a492e4dcd45f487638eff716c