set up cache busting based on app versions
noticed that the fonts were reloading on every pageload which is BAD.
but then the problem is how do you make sure the cache is stale when
you want it be to.

easy enough: use the version number in the url as a query param
for attachments.

also only set max-age header in prod. Otherwise we will be caching
development files unintentionally.

Sources:
- https://stackoverflow.com/a/22339262
- http://ankitjain.info/ankit/2013/12/05/cache-busting-nodejs-express-jade/