update node-sqlite library
for some reason, the node-sqlite library updated itself and broke
all the code that was working fine and dandy before.

but whatever, I know how to roll with the punches.

changed all of my code so that it does the exact same thing but
in a different way.

pretty sick. went back to this tutorial for help.
- https://medium.com/@tarkus/node-js-and-sqlite-for-rapid-prototyping-bc9cf1f26f10