decompose "getISODate" logic into two public functions
we want to be able to get the literal holiday date as well as the observed
date. currently, "getISODate" will only give us the observed date.
Adding an if condition would only make the logic harder to understand,
so I created two functions: "getLiteralDate" and "getObservedDate".

Currently still using getISODate in the code, this was just to
create the new functions and then test that they give the
expected output.