remove getISODay entirely
it did lots of good work for a long time.

however, we now want to return the literal date as well as the
"observed" date. `getISODay` didn't let us do both, so I broke the
logic into three functions.

Not the prettiest thing you've ever seen but all the tests pass.