change 'date' key to 'observedDate' key
the idea here is that the observedDate accounts for weekends
and tries to find mondays. the "Date" proper is the literal date of the
holiday.

So, for example, if Boxing Day is on a Saturday, it is observed
the following Monday.

The "date" would be the 26th while the "observedDate" would be the 28th.