Add "date" key to holidays and update Swagger spec
Added back the "date" key as the literal date of the holiday.

Makes sense, as this seems like the literal meaning of the word "date".

Updated and version bumped the swagger spec with the changes.

I've decided it's not a breaking change because your code won't be
broken, although I have slightly modified the format of the data.
If I had more users, I would consider it a breakign change, but I
don't really.