update API documentation online and on GitHub
basically, removed the markdown stuff that was on github.
Since the source of truth is already split between the docs and
the the actual Working Code, having _another place_ to keep up to
date just didn't make any sense.

Not sure if anyone was using those docs, but the swagger stuff
is more detailed.

Kept the brief overview with a listing of the routes and the accepted
query parameters. If you read that, you'll have a pretty good idea
how to use the API, so seemed worth keeping.