redirect remove
because heroku doesn't give you server control, you have to do a lot
of things at the level of the application.

redirecting from http to https was breaking on google cloud run where
they seem to internally be routing through the http protocol through
a reverse proxy or something.

removing this makes the app work on cloud run, basically.
otherwise, we're in a redirect loop.