add a cloudbuild.yaml config file
passing variables through `gcloud builds submit` command doesn't seem
possible unless we use the "--config" file alternative.

put all of the commands needed to build, push, and deploy the app into the
yaml config and now we're all good in the hood.

had to do some manual finagling to get the service account I created to
have the right permissions, but that should be addressed in the previous
pull request.

the best way I found to debug this was signing in as the service account:
- https://medium.com/@pnatraj/how-to-run-gcloud-command-line-using-a-service-account-f39043d515b9

and the docs on adding new variables are below:
- https://cloud.google.com/cloud-build/docs/configuring-builds/substitute-variable-values