Breaking change: 3.0.0
Moved all the /province URLs to /provinces
- eg, /province/MB/2021 is now /provinces/MB/2021

Looked up the docs on this and it looks like it's fine if you just
make sure you have 301s everywhere you used to have a link.

Added a bunch of tests and updated the sitemap as well so I think
we're basically in the clear here.

Normal users shouldn't notice a difference, nor will API users.

Sources:
- https://developers.google.com/search/docs/advanced/crawling/site-move-with-url-changes
- https://www.semrush.com/blog/canonical-url-guide/