Bugfix: Check holiday exists before adding to a province's list of holidays
Holidays returned from the DB are getting removed if they are before
their "firstOccurence" or after their "lastOccurence" dates.

But the "ProvinceHoliday" table has no idea about this, it just returns
a list of province/holiday combos.

What that meant was that there was a bunch of provinceholidays that were
like "you should add this Mourning Day to BC", but there was no longer
a Mourning Day in the list of holidays, so an "undefined" got pushed into
BC's array of holidays.

Then, later, we would loop through holidays looking for the name or the
date or something, and it would throw a 500 error because we were trying
to access a property of undefined.

Essentially, all prior or future years would fail for provinces with
a holiday that was in the future or in the past.

Big yikes.