Rework the h1
Previously, I had a separate <span> with a visuallyHidden human-readable
sentence, for if a screen reader or a Speakable interface was
wanting to read the first header of the page.

The main thing that I was doing was to change the order of the items in the header.

Now, I am just using CSS instead so I don't need all that extra markup.
Might look weird on my nintendo DS but most browsers now understand
display: flex.

Changed a bunch of tests now that the heading markup changed.