Update the status of National Truth and Reconciliation Day for Nunavut
It is a public service holiday, not a day everyone gets.
Here's a link: https://gov.nu.ca/justice/news/national-day-truth-and-reconciliation

It's a bit sneaky, but it says it applies to public service employees.
Means that it doesn't apply to everyone.

Also updated some of the stuff in the OpenAPI file where it had
some outdated docs about only AB and BC having optional holidays.