README: adjust getCurrentHolidayYear so that app doesn't break
Basically, getCurrentHolidayYear finds the latest holiday for the
region and adds 1 day (minus 1 minute) to it.

It also gets the current datetime.

If the current datetime is _after_ the last holiday day of the year,
then it will assume it's the next year.

On boxing day specifically, this fucks everything up because it
will say that today is before boxing day but then it won't have any
more holidays for this year.

If it thinks it's still 2023, then boxing day has no next holiday.

By making the addMinutes function return less minutes, we can make
the time before it should be, if that makes sense, and get around the bug.

As I'm writing this, I'm wondering if I should actually subtract minutes
from the boxing day date, but I'm not sure.

I'll fix it and then next year me can puzzle it out.