Manually set 'timestamp' when creating ICS files
Otherwise all my tests blow up.

Timestamp is not actually a documented value for the ics events, but
if you don't put something for it, it creates a string for me and then
complains that the string is the wrong format.

Wrote a small method that returns a unix timestamp for the time the
event is created and that seems to work.