Clean up API documentation of query params
Previously I was using the vertical bar to literally mean "OR".

eg: `?federal=true|1|false|0`

The intention was to communicate that either of these values were
accepted, not that they could be combined. The problem with the years
in particular is that sometimes people want more than 1 year.

(so like `?year=2024|2025`).

However, this doesn't work. `year` can only be 1 integer.

I've had a few people be confused by this, so now I have changed the
documentation to clarify that only 1 year is possible at a time.