Make selectable years dynamic
Instead of showing a fixed range of years, we will show people 4 years
ahead and 2 years behind.

Unless, there are less than 4 years to the max year we accept
(eg, if we are viewing 2032, we will only show 2033, no 2034),
and then we only show the remaining number of years.

Same thing if you are going backwards.