Add Jan 2 holiday for Quebec
Someone reached out telling me that Quebec celebrates "Traditional
holiday" on January 2nd, but it is for banks and public servants.

When you look for it, you do, in fact, find it to be there:

- https://www.quebec.ca/gouvernement/travailler-gouvernement/travailler-fonction-publique/services-employes-etat/conditions-travail/conges-vacances-annuelles
- https://www.nbc.ca/contact-us/bank-holiday-calendar.html
- https://www.banqueducanada.ca/medias/evenements-a-venir/calendrier-jours-feries-banque-du-canada/
- https://td.intelliresponse.com/cbaw/index.jsp?requestType=NormalRequest&source=100&id=6106&question=What+is+TD%27s+Holiday+Schedule

The other difference I noticed is that in Quebec, they just call it
"the day after new years", although apparently in English it is sometimes
referred to as "Traditional Holiday".

So to be consistent, I figured I would call it:
- EN: Day after New Year’s
- FR: Lendemain du Jour de l'An

This fits more or less into my category for "optional" holidays,
so I figured I could just add it there.